make animation times usable again
--durationTime that was used for the animations
was an integer. This was created extremely fast
animations for x1 and extremely slow animations
for x3. This fix changes that by setting the
durationTime in the qml side to real type. And
by that we are setting the following pragramatic
speeds which create a very good experience.
  x1 = 1.65
  x2 = 2
  x3 = 2.35
the above can be updated a bit based on the user
feedback and experience