fix #574,fix internal identification of tasks
--as in an older commit has been documented the qml
function childAt doesnt work propery for our ListView,
there are cases that returns an empty/null child. For this
reason we have implemented our own childAtPos function in
order to immulate the save behavior. Unfortunately there
wasnt a check in it that the items checked based on their
geometry they were actually TaskDelegate's. This fix not
only fixes bug #574 but also should provide a cleaner dragging
behavior, at least for the dragging phase