use the lastValidSourceName to paint the icon
--during removing a TaskDelegate the window is destroyed
in some cases and the icon provided by libtaskmanager
is a unknown pixmap. In such cases we should paint the
last valid one. That way it is always painted the correct
svg file even when removing a window that doesnt have
a launcher and the libtaskmanager provides an unknown
pixmap for it