fix #961,Semi-Tranparent Panels at ImproveVisibility
--update issue #957 in order to take into account
semi-transparent panels. With semi-transparent panels
contrast could be a puzzle. The changes that
introduced with this patch are:
1. Enable Improve Visibility when also case [E] from
 #957 is enabled.
2. When a semi-transparent panel exists in combination
with Improve Visibility option then add a threshold
in order to indicate if the Colorizer will be applied
or not. The threshold is 40% in order to indicate if
a semi-transparent panel is solid or not.