Activate next task in group on shortcut
Clicking a task group shows a tooltip with list of tasks. This tooltip
cannot be controlled by keyboard and it will stay open unless mouse is
used.

This patch changes the behavior so that when a global shortcut activates
a task group (usually Win+1, Win+2 etc.) a next task in the group is
activated without showing a tooltip.