fix launch feedback for plasma 5.13
--in plasma 5.13 there are cases the libtaskmanager
shows two launchers because we delay the launcher
removal for the bouncing animation. A way
to fix this by the user is to disable all startups
through plasma systemsettings->launch feedback
for the taskmanager. But because this is a
default plasma option many users would feel
uncomfortable when the launching animation
breaks. Such apps were gimp and inkscape.
This patch handles that case in order to provide
a nice no-breaking launcher animation for
these apps also.