improve dock settings windows geometry calculation
--the dock settings window calculations should not look
different between different font settings/geometries/screens.
One way to overcome that issue is by using the font.height to
calculate the settings height, that is 37chars for now and
the width to be calculated in accordance with that height. There
are some down limits for both, height shouldnt overcome screens
height and shouldnt go under 400px  and the width can not be
less than 200px

CCBUG: 397038