add customBackground class for colorizer
--in order to overlay properly the panel background
a roundness option was needed that takes into account
which borders should be drawn. In an upcoming commit
the user will be able to adjut the current plasma
theme roundness. The Colorizer is using now the
CustomBackground in order to draw the background
when the user wants to draw the contents based on
the current active window color scheme.