add LICENSE and AUTHORS in icons folders
--Debian packagers requested these files
in order to confirm the creators and licenses
for icons. This has arised again
in the past when Latte was introduced
for Ubuntu so it is good to end this
once and for all.

The creator for logo icons in folder "icons/"
is Alexey Varfolomeev and has granted GPLv2
for its icons. Alexey is a well known kde artist
and Latte project is thanking him for its fantastic
contribution, we love its icons! :)

The creator for pattern icons in folder
"containment/package/contents/icons/" is
me, Michail Vourlakos and I have also granted
for them a GPLv2 license.

The creator for icons in folder
"shell/package/contents/images/" is me,
Michail Vourlakos and I have granted
for them an LGPLv2+ license because of
their simplicity

BUG: 400758