restore Fitt's Law for edge Applets
--when the Latte View is in Justify mode
and and an applet is touching the screen edge
then for that applet any margins from the
Screen edge should be removed. That way
the Fitt's Law is obeyed. What is going to
happen for the margins of these applets with
their neighbour applets will be considered
in the future.

BUG: 408412