enable normal Latte Tasks removal action
--requesting from users a different way to remove
the Latte Tasks plasmoid if they want to is not
a good hig. Reenabling the classic ways to remove
Latte Tasks. The user can now use the applet tooltip
to remove it or by right clicking it in Latte
edit mode. Because of this the Tasks->Recycling
button is not needed any more.

BUG:415274