improve SunkEvents implementation
--each origin mouse area is now informin
the SunkEvents handler for the relevant
origin areas at the same time and this way
there is no events breakage when changing
from one sunked event origin area to another
--the sunked events origin and destination areas
are now released only when the event is NOT inside
any origin area or any destination. This way
no events breakage is appearing when changing from
destination area to origin area and vice versa.