plasmadesk:fix startup freezes from QDbusInterface
--as it appears plasma applets were not the main faulter
to blame for startup freezes. The bug was that Latte
was trying to communicate with Plasma DBus interface
to inform it about docks/panels geometries. For some
reason during startup other applets could also were
trying the same and for that reason Qt provides a 25sec
forced delay in case there are too many such calls.
The new approach does not block the startup code at all
and it should work at all cases.

BUG:444739
FIXED-IN:0.10.4